setup: fix link issues
- Since .env is linked, don't overwrite it (and prevent issues with
  follow up builds).
- If link is already there, don't do anything.
- Comments.

+++ setup.sh
@@ -28,9 +28,10 @@ build()
 
 link()
 {
-	src="${1}"
-	dst="${2}"
+	src="${1}" ## source to link
+	dst="${2}" ## target directory, where to place the link
 	test -f "${src}" || die "file '${src}' is missing or not accessable"
+	test -L "${dst}/$( basename ${src} )" && return || true
 	cd "${dst}"
 	ln -s "${src}" .
 	cd -
@@ -100,8 +101,6 @@ if [ -n "${DO_BUILD}" ]; then
 	mv ${TOPDIR}/${DOWNLOADDIR}/Xilinx_Unified_${VERSION}_*_Lin64.bin "${TOPDIR}/${DOCKERDIR}/build_context"
 	build "${DOCKERDIR}" "${DRYRUN}"
 	cd "${TOPDIR}/${DOCKERDIR}"
-	echo "!!! Docker finished, overwrite ${DOCKERDIR}/.env file with default user, in case adjust manually !!!"
-	do_env
 fi
 
 echo "READY."
